[Sum] - createBox func added + html + css
Started with doing the sum problem, added a function called createBox() in common, then added all the css and html to the first folder. Also added a function called "handleBox" which will be different in each page. (in this page it will work as a function to mark the one you click on and adding a "target" css class)