[Sum] +Handle sum of boxes (target & all boxes)
Fast changes to html and css, making more elements..

Codewise i added a sum function that will work with both the target boxes textContent and also all the boxes that you create from the start, the function takes in an arg which is a list of all the elements with the class. The function will then run and loop thru each element that was in the querySelectorAll, adding the number in the textContent to a variable and then return the sum of all these, which will then be displayed in a div.

//The function then gets called both when you click on a box (when you add a target class) and also each time you generate new boxes