[Reps] Repeated nums with colors
Added a new value in the returned object from mostRepNums, this will also send the doms that we can find and not only the textcontent. This makes it easier to put a "repeated" class with color- (i need to have this outside of the if statement in line 24, because i need to select all of the same numbers, not only the first one). + i added an easy if statement to check if nothing repeats and changed the textcontent as well in the border box