[Reps] find missing numbers func
I added a function to find all the missing numbers from 0 to 100 that aren’t assigned to a box. It loops through each number and checks if it’s in any box’s textContent. If it finds the number, it sets a variable "found" to true and breaks the inner loop. If not found, it adds the number to an array. Once all numbers are checked, it returns the array of missing numbers, which handleCreation() displays in the DOM.

I also noticed that the border-text-box were bigger in width compared to the other pages, which i added css for