[Rem] +markNum(), marks boxes with same textContent as the random num
Added a markNum function that takes in the random number, it will also loop thru all the boxes in the DOM and then first remove all the previously marked (to clear all the orange boxes each time you get a new random number). Then it checks with an if-statement if the textContent of the box  is the same as the random number, if it is then we can add the class "marked", this class has the background color orange in the css file.

in the global scope i added an eventListener to the button when you click it; this will then set the randomNumber variable to a random number by using randomNum() then it will update the textContent and run the function markNum(randomNumber) and then also "clear" the textContent for the removal-count box.