[Rem] +removeNum(), checks marked, adds new backgroundColor and textContent to each.
added removeNum() that will first of all check if there are any elements with the .marked class, if there isn't any that means that the markNum() "failed" (it couldnt find any elements with the .box class that had the same textContent as the random number). This will then put out "nothing to remove" as text. But if its longer than 0, it will loop thru all the elements with the .marked class and then change their backgroundColor to "tomato" and textContent to "X" to show that they have been removed. After the for loop it will display in a box. The randomNumber will be the value that i set it to once i have clicked the New Random Number button. I will also display how many times the number has been removed, i do this by checking the length of how many elements that have the .marked class.

In the global scope i will also add an eventListener to the Remove button which will just simply run the function.