new func in common.js that marks numbers with colors in the grid
Made a new function that both FindSame and Remove uses. I noticed that they both need a function to mark the numbers on the grid with colors, since they are exactly the same i decided to move it to common.js and then use the function from there in both pages but different arguments.