[Clear] handleBox() adds an eventListener to each that will mark target if clicked
Added css for the colors on hover and when clicked, also added a handleBox function that adds an eventlistener to each box, when you click on a specific box it will first check if it already has a target class in the classList, if it does it removes the class and sets the textContent to a random number (the function that is already set in common.js). Else if it doesnt contain target, it will add the target class and set the textContent to an empty string (it will be shown as if theres no text)

i also added handleCreation since it needs to exsist to run the code, as i explained before. Because it will be undefined.