[add] addsUp() func that will add a class if it finds boxes that adds up to the param number
Added a function called addsUp(sumNum) that wil take in the number thats in the input, first it will set a variable called foundCell to false, this is later used to break the loop so it doesnt continue checking after numbers, the first one it finds gets marked and not every number it finds. It will loop thru boxes in a nested loop to check each box and compare them to eachother. if its the same box however it will continue, skipping it so it doesnt add up the same box, it can be the same number, but not the same DOM element. it will then check if the textContent of the box in the outer loop is the same as the inner loop (important: this needs to be parsed so we are looking at an int instead of a string since we are adding them up, if its two strings it will count as an concatination) i will then check if the sum of these two boxes is the same number as the one in the input. If its true, it will add a class to both boxes, set the foundCell as true, and then break. It will also check in the outer loop if its true and then break there as well. The reason for this is because  also want to break from the outer loop since it can continue finding matching numbers there.