[add] button to find the cells
added a button with an eventListener if you click it, it will first loop thru the boxes and check if it already has the class, if it does, remove it from each box with the class. then in the end it will call the function adsUp with an arg that is the value of the input.

something intressting i found here was that both clear and addUpTo has the same kind of button, i need to check if the class already exsists somehow. I will probably put this in common.js instead since the code is repeating