removeClassFromAll() func in common.js that removes marked to reset
I added a new function, i found that the clear, addUpTo and findSame pages uses the same kind of loop and removes a class to reset all the classes on the boxes. I decided to make this into one function in common.js instead. now it will take in an arg, otherBoxmodifyFunc and see firstly if its true. This is because of the clear page; it uses a new random number as text content if you click the button on the page. So if its not true, i didnt put any args so it will be undefined and therefor skip it. Else if it is true and defined i will run a function which i have in clear.js  ((box) => { box.textContent = randomNum(0, 100) }) this will change a box textContent to the random num, and i will also in this function put box as an arg which is boxes[i], as i set it in common.js. Therefor i can take the box i am on in the loop and just change it textContent with a function as a param to removeClassFromAll. It will work on the other pages as well because of the check i have if its true or not, if its not its undefined and wont do anything. If it runs the function with boxes[i] as a param that is the current box i am on and will change its textContent.